test: remove duplicate slugify and genSecret tests
